FIX: reorganize tests
