+++ tests/image/data/48by48.png
