+++ tests/image/data/48by48rgb.png
