+++ tests/image/data/48by48rgba.png
