+++ tests/image/data/64by64.png
