+++ tests/image/data/happy.png
