+++ tests/image/data/keyboard.jpg
