+++ tests/text/test_organizations.py
@@ -1,15 +1,9 @@
-import unittest
 
-from indicoio import config
+
 from indicoio import organizations
+from indico_text_base import TextTest
 
-
-class Organizations(unittest.TestCase):
-    def setUp(self):
-        self.api_key = config.api_key
-
-        if not self.api_key:
-            raise unittest.SkipTest
+class OrganizationsTest(TextTest):
 
     def test_organizations_v2(self):
         test_data = "A year ago, the New York Times published confidential comments about ISIS' ideology by Major General Michael K. Nagata, then U.S. Special Operations commander in the Middle East."
